requirements.txt
torch>=1.10.0
torchvision>=0.11.0
transformers>=4.15.0
timm>=0.5.4
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0
Pillow>=8.3.0
opencv-python>=4.5.0
tqdm>=4.62.0
tensorboard>=2.7.0
albumentations>=1.1.0
jupyter>=1.0.0
notebook>=6.4.0
ipywidgets>=7.6.0
plotly>=5.5.0
